refactor: Response 클래스 범용적으로 설계 및 변경
