+++ src/main/java/controller/HomeController.java
@@ -3,6 +3,7 @@ package controller;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import webserver.HttpStatus;
 import webserver.Request;
 import webserver.Response;
 
@@ -20,8 +21,6 @@ public class HomeController implements Controller {
 	@Override
 	public void process(Request request, Response response) throws IOException {
 		byte[] body = Files.readAllBytes(new File("./webapp" + request.getPath()).toPath());
-		response.response200Header(body);
-		response.responseBody(body);
+		response.write(body, HttpStatus.OK);
 	}
-
 }
