+++ src/main/java/controller/UserLogoutController.java
@@ -1,6 +1,9 @@
 package controller;
 
-import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import util.Pair;
+import webserver.HttpStatus;
 import webserver.Request;
 import webserver.Response;
 
@@ -17,7 +20,10 @@ public class UserLogoutController implements Controller {
 
 	@Override
 	public void process(Request request, Response response) {
-		response.logoutHeader("http://localhost:8080/index.html");
+		List<Pair> pairs = new ArrayList<>();
+		pairs.add(new Pair("Location", "http://localhost:8080/index.html"));
+		pairs.add(new Pair("Set-Cookie", "sessionId=; max-age=-1; Path=/"));
+		response.write(HttpStatus.FOUND, pairs);
 	}
 
 }
