+++ src/main/java/util/IOUtils.java
@@ -4,7 +4,6 @@ import java.io.BufferedReader;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
-import util.HttpRequestUtils.Pair;
 
 public class IOUtils {
     /**
