+++ src/main/java/webserver/Request.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import util.HttpRequestUtils;
-import util.HttpRequestUtils.Pair;
+import util.Pair;
 import util.IOUtils;
 
 public class Request {
